requirements.txt
aiofile==3.8.8
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
caio==0.9.13
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpath-ng==1.6.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37.post1
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.19
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
loguru==0.7.2
lxml==5.2.1
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnxruntime==1.18.0
openai==1.28.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
ply==3.11
portalocker==2.8.2
preshed==3.0.9
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
streamlit==1.35.0
streamlit-chat==0.1.1
striprtf==0.0.26
sympy==1.12
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
transformers==4.40.2
triton==2.3.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
whatsapp-python==3.0.6
wrapt==1.16.0
yarl==1.9.4
